[LuxMenuBar] refactor deep css in LuxLibraryHeader into a MenuBar theme
Prior to this commit, LuxLibraryHeader edited the
colors of child LuxMenuBar components according
to the value of its theme prop.

This commit moves this CSS to LuxMenuBar, under
the control of its own theme prop. This has a
few benefits:

* Better separation of concerns (LuxMenuBar is
  now responsible for its own CSS, rather than
  LuxLibraryHeader taking responsibility for
  LuxMenuBar's CSS in certain scenarios).
* These themes are now available to use even
  if LuxMenuBar is not a child component of
  LuxLibraryHeader

There is also a tradeoff: if you pass a LuxMenuBar
into the LuxLibraryHeader slot and want to use the
light theme or shade theme, you will have to specify
the theme on both components, instead of just on
LuxLibraryHeader.  I could find no usage of the light
or shade theme in the pulibrary org, so I consider
this to be an acceptable tradeoff.

Many thanks to @christinach for figuring this out
with me!

Helps with #216 and pulibrary/allsearch_frontend#296