[luxUniversityAccessibility] show that the link opens in a new tab
This also adds an SVG icon for opening in a new tab, and adds a prop to LuxHyperlink
to toggle the icon on or off.  When the SVG icon is shown, its accessible name is
set to 'opens in new tab', to let screen reader users also know about it.

Closes #265