[LuxIconBase] use valid ids for connecting <svg> to its accessible name in <title>
IDs should not contain whitespace.  Consider the following pair of tags, which
were generated prior to this commit:
<svg aria-labelledby="(opens in new tab)">
<title id="(opens in new tab)">

If there are spaces in the aria-labelledby, it is supposed to treat it as a
space-delimited list of ids, so look for an element with id="(opens", one with
id="in", and so on.

In practice, Voiceover + Safari was smart enough to know that we meant these
two elements to be linked, but there's no need to risk it.  This commit removes
spaces and other potentially problematic characters from the iconName prop
before using it in `id` or `aria-labelledby`.