[LuxMenuBar] Add new unsafe_name property to the menu item object
`unsafe_name` works like the existing name property, except that if it contains some
arbitrary HTML, it does not escape it.  The immediate use case for this behavior
is the "Bookmarks (4)" menu item in the catalog, which has a span with a specific class
around the number that allows Blacklight's javascript to update it as the user adds/removes
bookmarks.

This has security implications, since `unsafe_name` bypasses Vue's cross-site scripting
protections.  I've added documentation about this (as well as other properties on this
object that did not yet have this documentation).

Interestingly, the `buttons` version of LuxMenuBar was already bypassing cross-site
scripting protections for the `name` property.  In the interest of consistency and safe
defaults, this commit make it match the behavior of the others: you can add arbitrary
HTML with `unsafe_name`, not with `name`.

Helps with pulibrary/orangelight#4088