Improve recommendations for adding Lux to a static HTML page
When we first added Lux to LibCal, we mounted it to the page's body,
which conflicted with some other javascript that was necessary to
successfully make appointments (the appointments would go through,
but the user would get confusing error messages and unstyled content).

We found that this issue goes away if you mount separate areas of
the DOM as separate Vue apps.