[LuxAlert] Add a prop to optionally configure how long before it autoclears
By default, if an alert is set to autoclear, it disappears after 2 seconds,
which is a very short amount of time to read certain messages!

This commit adds a new prop called autoclearSeconds where you can optionally
set this to a higher number.  The use case for this is in the catalog, so
we can replace the blacklight/bootstrap alert flash message (which waits
5 seconds before disappearing) with LuxAlert.

Also, refactor the autoclear tests to use jest fake timers, so they don't
have to wait all those seconds before running their assertions.