[LuxMenuBar] If no href is supplied, use a <button> tag instead of an <a> tag
Prior to this commit, if you don't supply an href value for
a child menu item, Lux will render an <a> tag without an
href.  This is not focusable, so keyboard users will not be
able to activate that menu item.

This helps with https://github.com/pulibrary/allsearch_frontend/issues/443,
since it allows us to create menu items that are keyboard
focusable that are not links (in this case, they will be
buttons that switch between dark mode and light mode).