[#459] Don't cut off the edges of arrow icons
LuxArrowUp and LuxArrowDown both have a stroke width of 3.
When the stroke width is just 1, these icons do fit within
the viewbox provided in LuxIconBase.  However, the larger
stroke width pushes them outside the viewBox.

Here, we use the overflow: visible attribute to prevent the
edges of these icons from being clipped.

Thanks very much to @christinach for investigating this
issue with me.

Closes #459

Co-authored-by: Christina Chortaria <christinach@users.noreply.github.com>