[LuxShowMore] Don't display HTML comments as part of the snippet view
The previous implementation of fullText() recursively traversed the
nodes in the slots, looking for any text that might be included in the
snippet.  Unfortunately, this approach did not exclude HTML comments (which
appeared in the snippet view though they should not), and failed to
include text provided by v-html bindings (which should be included).

This approach instead keeps the full text in the DOM, even before
the user has pressed the ShowMore button.  When the LuxShowMore is
not expanded, it is hidden using display:none, but we can still call
textContent on it to retrieve text -- including v-html bindings and
excluding html comments -- to construct the snippet.

Co-authored-by: Christina Chortaria <christinach@users.noreply.github.com>
Co-authored-by: Max Kadel <maxkadel@users.noreply.github.com>
Co-authored-by: Ryan Jensen <RyanAJensen@users.noreply.github.com>
Co-authored-by: Ryan Laddusaw <rladdusaw@users.noreply.github.com>