[LuxAutocompleteInput] Announce the current selection when navigating using arrow keys
Prior to this commit, if you:
1. Open Mac VoiceOver
2. Go to the styleguide demo for the LuxAutocompleteInput component
3. Try navigating through the fruits using your arrow keys

VoiceOver will be completely silent as you move up and down, other
than the occassional "bump" sound effect as your cursor moves around
in the input field.

This commit uses aria-activedescendant to let voice over know which
item to announce as the user moves around through the list of fruits.
It changes the role of the input to combobox (aria-activedescendant is
only valid on certain roles, combobox is one of the valid roles).
It also prevents the default behavior of the up and down arrows (moving
to the start and end of the input), since that causes an annoying bumping
sound that interrupts all the fruit announcements from the autocomplete
entries.

I have also confirmed this fix on NVDA with Firefox.

Closes #352