Remove repeated definitions of CSS variables from built CSS
Prior to this commit, we imported the CSS variables into each component that needed them.
This allowed the components to display with the correct CSS, but came at a cost: in the
built CSS, the CSS variable definitions were repeated 24 times, once for each component
that used them.

This commit takes a different approach: we do not import the CSS variables into each
component, but rather load the variables once.  For the styleguide display, this
takes place as part of the styleguide configuration.  For the build CSS,
this happens via an import in index.js.

This reduces the size of the compiled CSS from 234 kB to 128 kB.